+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 3.x
 
+## Improvements
+
+- DAPI-19: Add filter on subscribed/unsubscribed products to Franklin Insights in the product grid
+
 ## Bug fixes
 
 ## Technical improvement